third_party/WebKit/LayoutTests/fast/parser/parse-menuitem.html

<!DOCTYPE html>
<head>
<script src="../../resources/js-test.js"></script>
</head>
<div id="div"><menuitem></div>
<menu id="menu" type=popup><menuitem id="first_item"><menuitem id="second_item"></menu>
<script>
description("This test checks parsing of menuitem tag which is a self closing tag with no end tag.");
shouldBeEqualToString("document.getElementById('div').innerHTML", "<menuitem>");
shouldBeEqualToString("document.getElementById('menu').innerHTML", "<menuitem id=\"first_item\"><menuitem id=\"second_item\">");
shouldBeNull("document.getElementById('first_item').firstChild");
shouldBe("document.getElementById('menu').firstChild.nextSibling", "document.getElementById('second_item')");
</script>
